Fix bb shell example function in README
Right now running `bb` will not return the currently playing song as described in the readme, but the usage info.

This patch fixed that. Displaying the current song is now possible and the usage is only show if an unknown command is run (e.g. when running `bb h` or `bb help` :wink:).